A. Decimal To Binary

+++ Yandex 5.2/A.DecimalToBinary.cpp
@@ -21,7 +21,7 @@ void decToBinary(long long n)
         cout << binaryNum[j];
 }
 
-int decimalToBinary(long long n)
+long long decimalToBinary(long long n)
 {
  
     // To store the binary number
